+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KRGST59BE6P7KM26YJTMBRKW"
+        "01KRK06RSYE98BAZ82RZ66WWND"
     )
     r"""The unique identifier of the data source"""